Webhook splitting into field and record updates
